Fix the bit-field related compile warnings
There are 2 bit-fields defined as 'int' with structure 'pcm' and
later instant numbers 0 and 1 are assigned to them. Accordingly,
some compile warnings are caused which complain that the implicit
truncation changes value from 1 to -1. Actually the related code
is checking to see if the bit-fields are 0 or not.

With this patch, the definition of these 2 bit-fields have been
changed to 'bool'. The related value assignments have also been
adapted accordingly.

Test: mm libtinyalsa, presubmit check.

Change-Id: Ic036edd80296800744b217b6e2ac0ac0e00fb812
Signed-off-by: Jing Mike <jingyangliu@eswincomputing.com>